📝 Add error state and helper text to InputText component
This commit adds the `error` prop to the `InputText` component, which when set to true, displays the input field in an error state. It also adds a new prop called `helperText`, which is displayed below the input field and can be used to provide additional information about the error. The commit also includes changes to styles and tests related to this feature.