🎨 Add new button icon options and transparent background option to README
This commit adds two new optional props to the button component: `icon` and `iconOnly`. The `icon` prop allows for a custom icon to be displayed when the button is not in `iconOnly` mode. Additionally, a new prop called `transparent` has been added which allows for the button to have a transparent background.