📝 Added TextArea component
A new TextArea component has been added to the library. This versatile input field supports larger inputs and allows for customization and theming. The README.md file has been updated with a detailed description of the props that this component accepts, along with usage examples.

In addition, storybook stories have been created for the TextArea component demonstrating its use with different column and row configurations.