🔄 Added loading state to InputText component
In this update, a new 'loading' state has been added to the InputText component. This includes changes in the storybook and tests for the new feature. The loading state is visually represented by a LoadingWrapper that contains a Loading component. Additionally, some adjustments have been made to the styling of the input field when it's in its loading state. A 'size' prop was also introduced to control the size of the input field.